+++ main.go
@@ -16,6 +16,7 @@ func buildApp() *warg.App {
 			"Manage Environmental secrets centrally",
 			section.ExistingCommand("version", warg.VersionCommand()),
 			section.ExistingFlag("--color", warg.ColorFlag()),
+			section.ExistingCommand("init", cli.InitCmd()),
 			section.Section(
 				"env",
 				"Environment commands",